PID improvement: anti-windup + overflow hardening + D-filter fix; fw 0.15.3.4
Root causes fixed (all bench-confirmed, plan doc sections 1-18):
- kI>=10 oscillation was integrator-fueled stick-slip windup: integral
  authority clamped to 25% of output authority + back-calculation
  anti-windup at pid_output_limit (kI=25 post-move RMS 6,865 -> ~35)
- "D does nothing" was a sticky floor-decay bug: (lpf*31)>>5 never decays
  values -31..-1, leaving a kD-proportional D bias (-85k at kD=175k);
  decay now rounds toward zero; output shift rounds to nearest
- 15 UB integer-overflow sites hardened: error input clamp +/-2^29,
  authority cap 800e6, gains capped to INT32_MAX, ki capped for railed-
  integrator step safety, int64 recompute; proven by 3.24B-tick UBSan run
- Low-current-limit torque regression fixed: pid_output_limit =
  max(V<<8, 16384) so lead angle can reach 90 deg at V < 64
- No previous_error reset on constants change (derivative kick eliminated)
- D filter shift #ifndef-guarded; bench sweep confirms shift 5 optimal

Permanent verification harnesses (rerun after ANY PID change):
- c_test_programs/pid_algorithm_sim.c: headless firmware-exact sim with
  bench-calibrated stick-slip plant + ablation/study flags
- test_pid_differential.sh: verbatim firmware extraction vs reference
  model, bit-identity (6.5M samples clean)
- test_pid_overflow.{c,sh}: UBSan harness, all 3 product parameter sets
- run_pid_sim_battery.py (60/62; 2 documented pathological-plant corners)
- run_pid_sim_extended_sweep.sh (82/82)
- tools/pid_debug_capture.py: hardware capture via test mode 3 (~240 Hz)
  with built-in output-identity validation

Data: firmware/pid_baseline_captures/ (figs 1-11: baseline, sim, bench
verification, perf/assembly study, voltage tests, overshoot root cause,
D-filter study, ultracode pass, rounding ablation, load-holding study).
Verified: M17/M23/JIG builds green, mutation testing 16/16 kills, bench
campaign H1-H10 green, full hardware suite 63/63.

+++ c_test_programs/test_PID_controller.c
@@ -124,7 +124,7 @@ double PID_controller_double(double error)
 }
 
 
-int32_t PID_controller(int32_t error);
+int32_t PID_controller(int64_t error); // NOTE: signature changed in fw 0.15.3.2 (int64 input); this legacy harness predates the M17/M23-only product defines and does not currently build (see build_script.sh M3/M4 references)
 
 int main(void)
 {
